vda-sbt-client : Updated using @verida/vda-common
