+++ packages/vda-sbt-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+2023-04-12 (v2.3.0)
+-------------------
+
+- Updated using `@verida/vda-common`
+- Test code updated
+
 2023-03-27 (v2.2.1)
 -------------------
